refactor: move SecurePin to authentication-logic module
This commit relocates the `SecurePin` and `SecurePinImpl` classes, along with their associated tests, from `business-logic` to the `authentication-logic` module to better align with the project's architectural boundaries.

Key changes include:
- Moved `SecurePin` and `SecurePinImpl` to the `eu.europa.ec.authenticationlogic.model` package.
- Updated all references and imports across the `issuance-feature`, `common-feature`, `ui-logic`, and `test-feature` modules.
- Changed `CryptoController.generateCodeVerifier()` to a `suspend` function.
- Updated `ui-logic/build.gradle.kts` to include a dependency on the `AuthenticationLogic` module.
- Adjusted Kover exclusion rules for the `AuthenticationLogic` module to reflect the package changes.
