+++ authentication-logic/src/main/java/eu/europa/ec/authenticationlogic/controller/storage/PinStorageController.kt
@@ -17,7 +17,7 @@
 package eu.europa.ec.authenticationlogic.controller.storage
 
 import eu.europa.ec.authenticationlogic.config.StorageConfig
-import eu.europa.ec.businesslogic.model.SecurePin
+import eu.europa.ec.authenticationlogic.secure.SecurePin
 
 interface PinStorageController {
     suspend fun hasPin(): Boolean
