+++ issuance-feature/src/main/java/eu/europa/ec/issuancefeature/ui/code/DocumentOfferCodeViewModel.kt
@@ -18,7 +18,7 @@ package eu.europa.ec.issuancefeature.ui.code
 
 import android.content.Context
 import androidx.lifecycle.viewModelScope
-import eu.europa.ec.businesslogic.model.SecurePin
+import eu.europa.ec.authenticationlogic.secure.SecurePin
 import eu.europa.ec.commonfeature.config.IssuanceSuccessUiConfig
 import eu.europa.ec.commonfeature.config.OfferCodeUiConfig
 import eu.europa.ec.eudi.wallet.document.DocumentId
